temporal: CONVERT forces the evaluation order in targets
This change allows us to change the compiler to use targets
incrementally.

COMPILE-TOPLEVEL will emit first the code in the target. This can result
in code that is not executed in the right order. For example: the
primitive + could insert the type-check in the target, but because the
target is bound in COMPILE-TOPLEVEL and no other forms are using the
target, the type check will be executed too soon.

In order to support both returning expresssions and targets temporarily,
CONVERT will also bind *TARGET* and will return an immediately invoked
function expression (IIFE) composed of both the expression and the
target.

This should mostly work, as CONVERT is the main entry point to compile
all the forms.